@@ -2,10 +2,20 @@
   "name": "cli-kit",
   "version": "1.0.0",
   "description": "cli demo",
-  "main": "index.js",
+  "main": "lib/index.js",
+  "typings": "typings.d.ts",
+  "files": [
+    "lib",
+    "bin",
+    "typings.d.ts"
+  ],
   "license": "MIT",
   "bin": {
     "cli-kit": "bin/index.js"
+  },
+  "scripts": {
+    "build": "rimraf lib/* && ttsc",
+    "watch": "ttsc -w"
   },
   "dependencies": {
     "cac": "^6.5.10",
@@ -24,8 +34,21 @@
     "update-notifier": "^4.1.0"
   },
   "devDependencies": {
-    "eslint": "^7.2.0",
-    "eslint-plugin-node": "^11.1.0"
+    "@types/consolidate": "^0.14.0",
+    "@types/fs-extra": "^9.0.1",
+    "@types/metalsmith": "^2.3.0",
+    "@types/node": "^14.0.14",
+    "@types/semver": "^7.3.1",
+    "@types/update-notifier": "^4.1.0",
+    "@typescript-eslint/eslint-plugin": "^3.5.0",
+    "@typescript-eslint/parser": "^3.5.0",
+    "eslint": "^7.3.1",
+    "eslint-config-alloy": "^3.7.3",
+    "eslint-plugin-node": "^11.1.0",
+    "rimraf": "^3.0.2",
+    "ttypescript": "^1.5.10",
+    "typescript": "^3.9.5",
+    "typescript-transform-paths": "^1.1.14"
   },
   "engines": {
     "node": ">=8.10.0"
