@@ -1,14 +1,14 @@
-const message = require('./message');
-const download = require('./download');
-const generator = require('./generator');
-const {getAnswersTemplate} = require('../prompts');
+import message from '@/utils/message';
+import download from '@/utils/download';
+import generator from '@/utils/generator';
+import {getAnswersTemplate} from '@/prompts';
 
-module.exports = async function selectTemplate(toPath) {
+export default async function selectTemplate(toPath: string) {
   try {
     const answersTemplate = await getAnswersTemplate();
 
     // 下载模版，返回临时目录
-    const tmpDest = await download(`github:jsany/${answersTemplate.selectTemplate}`);
+    const tmpDest: string = await download(`github:jsany/${answersTemplate.selectTemplate}`);
 
     // 实现脚手架给模板插值的功能并复制到项目目录中
     await generator(tmpDest,toPath)
